consolidates Features and CustomFeaturesDefinition namespaces into Builders.Features
