+++ SolastaCommunityExpansion/Builders/Features/FeatureDefinitionConditionalPower.cs
@@ -1,7 +1,7 @@
 ﻿using SolastaModApi;
 using SolastaModApi.Extensions;
 
-namespace SolastaCommunityExpansion.CustomFeatureDefinitions
+namespace SolastaCommunityExpansion.Builders.Features
 {
     public interface IConditionalPower
     {
