+++ SolastaCommunityExpansion/Builders/Features/FeatureDefinitionRemoveGrantedFeature.cs
@@ -1,7 +1,7 @@
 ﻿using SolastaModApi;
 using SolastaModApi.Extensions;
 
-namespace SolastaCommunityExpansion.CustomFeatureDefinitions
+namespace SolastaCommunityExpansion.Builders.Features
 {
     //
     // RemoveGrantedFeature allows you to remove a previously granted feature from a hero character
