+++ SolastaCommunityExpansion/Classes/Witch.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using SolastaModApi;
 using SolastaModApi.BuilderHelpers;
 using SolastaModApi.Extensions;
-using SolastaCommunityExpansion.Features;
+using SolastaCommunityExpansion.Builders;
+using SolastaCommunityExpansion.Builders.Features;
 using SolastaCommunityExpansion.Helpers;
 using SolastaCommunityExpansion.Subclasses.Witch;
 using SolastaCommunityExpansion.Level20;
