+++ SolastaCommunityExpansion/Feats/FightingStlyeFeats.cs
@@ -1,4 +1,4 @@
-﻿using SolastaCommunityExpansion.Features;
+﻿using SolastaCommunityExpansion.Builders.Features;
 using SolastaCommunityExpansion.FightingStyles;
 using SolastaCommunityExpansion.Models;
 using SolastaModApi;
