+++ SolastaCommunityExpansion/Feats/OtherFeats.cs
@@ -1,4 +1,4 @@
-﻿using SolastaCommunityExpansion.Features;
+﻿using SolastaCommunityExpansion.Builders.Features;
 using SolastaModApi;
 using System;
 using System.Collections.Generic;
