+++ SolastaCommunityExpansion/Subclasses/Rogue/Thug.cs
@@ -1,4 +1,4 @@
-﻿using SolastaCommunityExpansion.CustomFeatureDefinitions;
+﻿using SolastaCommunityExpansion.Builders.Features;
 using SolastaModApi;
 using SolastaModApi.Extensions;
 using System;
