+++ SolastaCommunityExpansion/Subclasses/Witch/WhiteWitch.cs
@@ -1,4 +1,4 @@
-using SolastaCommunityExpansion.Features;
+using SolastaCommunityExpansion.Builders.Features;
 using SolastaModApi;
 using SolastaModApi.BuilderHelpers;
 using System;
